+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Pashmak Changelog
 
-## 0.6, 0.6-beta8
+## 0.6, 0.6-beta9
 
 #### New features
 - added **Class** system
@@ -17,6 +17,7 @@
 - added `fopen` alias for `open` function
 - added `py_load_file` function to load python scripts as object (GH-35)
 - added python `json` library support
+- added `function.{list,delete,exists}`, `class.{list,delete,exists}` super functions
 
 #### Bug Fixes
 - fixed a bug in `eval` function